+++ ensysmod/schemas/energy_commodity.py
@@ -3,7 +3,7 @@ from typing import Optional
 from pydantic import BaseModel, Field, validator
 
 from ensysmod.schemas import Dataset
-from ensysmod.util import validators
+from ensysmod.utils import validators
 
 
 class EnergyCommodityBase(BaseModel):
