+++ ensysmod/schemas/energy_storage.py
@@ -4,8 +4,13 @@ from pydantic import BaseModel, Field
 from pydantic.class_validators import validator
 
 from ensysmod.model import EnergyComponentType
-from ensysmod.schemas import EnergyComponentCreate, EnergyComponent, EnergyComponentUpdate, EnergyCommodity
-from ensysmod.util import validators
+from ensysmod.schemas import (
+    EnergyCommodity,
+    EnergyComponent,
+    EnergyComponentCreate,
+    EnergyComponentUpdate,
+)
+from ensysmod.utils import validators
 
 
 class EnergyStorageBase(BaseModel):
