+++ ensysmod/schemas/region.py
@@ -1,9 +1,9 @@
 from typing import Optional
 
-from pydantic import BaseModel, validator, Field
+from pydantic import BaseModel, Field, validator
 
 from ensysmod.schemas import Dataset
-from ensysmod.util import validators
+from ensysmod.utils import validators
 
 
 class RegionBase(BaseModel):
